updated to new completed approach to retrieving list items.
must uncomment the sorting functionalities from db file and servlet
file. shouldn't be a problem.

after seeing how java and requests handle arrays being sent over HTTP
(no indexing), we should probably leave the sorting inside the servlet
intact because i dont *think* the order is guaranteed, but it's
definitely something we can look into the documentation and see.
documentation: stackoverflow

not sure if every DB item is being retrieved, could be a DB issue (have
to paginate or use a projection expression to get less shit back) or it
might be a JS issue.

by the way, the way its set up now, the isFinal variable is set by JS,
because only the front end knows when the end is based on how many items
have been delivered. 

cheers